Enhanced front-page design
Still some issues such as inability for the countdown clock to work in pure js. The file is time.js in assets/js/time.js. I tried to load it but the error shows unable to get a value. I guess the problem could be due to the JS working differently after the React Webpack compilation: Var does not work as well?